+++ src/index.ts
@@ -1,0 +1,21 @@
+import {
+  ActionCreator,
+  action, payload,
+} from './action';
+import {
+  union,
+} from './actiontype';
+import {
+  SingleActionReducer,
+  reducer, subreducer,
+} from './reducer';
+
+export {
+  ActionCreator,
+  SingleActionReducer,
+  action,
+  payload,
+  reducer,
+  subreducer,
+  union,
+};